+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 * Add support for the `onAnimationStart`, `onAnimationIteration` and `onAnimationEnd` event handlers.
+* Add `animationDuration` and `animationTimingFunction` to `CssProps`.
 
 ### [v0.170](https://github.com/react4j/react4j/tree/v0.170) (2020-07-20) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.169...v0.170)